Set up home page to show categories and apps
