+++ src/app/page.tsx
@@ -1,3 +1,7 @@
+'use client';
+
+import AppPage from './components/AppPage';
+
 export default function Home() {
-	return <></>;
+	return <AppPage edit={false} />;
 }
