+++ src/app/scripts/appsManifest.ts
